refactor(coordinator/api): Phase 6 — split the large handler files into per-concern files
Pure-move splits (deletion-only on each source; no body/comment/signature changed):
  model_registry_handlers.go 884 -> 9    (+ catalog/write handlers, auth, validate,
                                           manifest, projection)  — all 7 SyncModelCatalog
                                           calls + manifest verify ordering preserved
  billing_handlers.go        ~830 -> 21   (+ stripe/referral/pricing/admin/earnings
                                           handlers) — webhook idempotency order,
                                           un-auth'd http.Error webhook preserved
  stripe_payouts.go          ~690 -> 31   (+ onboard/withdraw/webhook/status/helpers)
                                           — withdraw ordering + asymmetric refunds +
                                           wd-tr-/wd-po- idempotency keys preserved
  me_handlers.go             ~640 -> 110  (+ types/fleet/identity/build_provider/
                                           enrich/counts) — buildMyProvider
                                           unlock-before-PendingCount preserved

gofmt clean, go build/vet, and go test ./coordinator/api/... (throwaway Postgres)
all green; each source file diff is deletions-only.
